changed how plugins are defined, added test and README
