+++ tests/integration/components/md-text-test.js
@@ -47,3 +47,45 @@ test('it renders a dynamic template', function(assert) {
   this.render(hbs`{{md-text text=tpl html=true dynamic=true}}`);
   assert.equal(this.$().find('a').length, 2); // should have both dynamic and static link
 });
+
+test('it renders a video with plugin', function(assert) {
+  let reg = /^%\[([^\]]*)\]\s*\(([^)]+)\)/;
+
+  this.plugins = [
+    function video(md, options) {
+      md.inline.ruler.push('video', function (state, silent) {
+        // it is surely not our rule, so we could stop early
+        if (state.src[state.pos] !== '%') {
+          return false;
+        }
+
+        var match = reg.exec(state.src.slice(state.pos));
+        if (!match) {
+          return false;
+        }
+
+        // in silent mode it shouldn't output any tokens or modify pending
+        if (!silent) {
+          state.push({
+            type: 'video',
+            title: match[1],
+            src: match[2].replace('/watch?v=', '/embed/'),
+            level: state.level,
+          });
+        }
+
+        // every rule should set state.pos to a position after token's contents
+        state.pos += match[0].length;
+
+        return true;
+      }, options);
+      
+      md.renderer.rules.video = function (tokens, idx/*, options, env*/) {
+        return `<iframe width="560" height="315" src="${tokens[idx].src}" frameborder="0"></iframe>`;
+      };
+    }
+  ];
+
+  this.render(hbs`{{md-text text='%[Test video](https://www.youtube.com/watch?v=OInJBwS8VDQ)' plugins=plugins}}`);
+  assert.equal(this.$().find('iframe').length, 1);
+});
